backend/requirements.txt
django==4.2
djangorestframework==3.15.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
